requirements.txt
# Core dependencies
Flask==2.2.2
influxdb==5.3.1
astral==3.2
pytz==2022.7.1
numpy==1.24.3

# Raspberry Pi GPIO and sensor libraries
RPi.GPIO==0.7.1
dht11==0.1.0

# Adafruit libraries for sensors
Adafruit-Blinka==8.39.1
adafruit-circuitpython-ads1x15==2.2.25
adafruit-circuitpython-busdevice==5.2.9
Adafruit-GPIO==1.0.4
Adafruit-PlatformDetect==3.63.0
Adafruit-PureIO==1.1.11

# HTTP and web dependencies
requests==2.31.0
urllib3==1.26.16

# Utilities
python-dateutil==2.8.2
